backend/requirements.txt
flask
pymongo
python-dotenv
requests
flask-cors
waitress
gunicorn
flask-session
qdrant-client
sentence-transformers
groq
